remove landing page
