+++ src/routes/(app)/auth/+page.svelte
@@ -9,7 +9,7 @@
 	async function verifyGoogleLogin(request: { credential: string }): Promise<void> {
 		try {
 			await fetchUser(request.credential);
-			goto('/portal');
+			goto('/');
 		} catch (e) {
 			notificationStore.set({
 				title: 'Cannot sign in',
@@ -22,7 +22,7 @@
 
 	onMount(() => {
 		if ($userStore !== undefined) {
-			goto('/portal');
+			goto('/');
 		}
 
 		if (google) {
