--- src/routes/(app)/portal/+layout.svelte
@@ -1,8 +1,0 @@
-<script>
-	import Fullpage from '$lib/components/layouts/Fullpage.svelte';
-	import './styles.scss';
-</script>
-
-<Fullpage>
-	<slot />
-</Fullpage>
