--- src/routes/(app)/portal/+layout.ts
@@ -1,21 +1,0 @@
-import { token, fetchUser } from '../../../stores/auth';
-import { redirect } from '@sveltejs/kit';
-
-export const load = async () => {
-	let userToken;
-	token.subscribe((value) => {
-		userToken = value;
-	});
-	if (null === userToken || undefined === userToken || '' === userToken) {
-		throw redirect(307, '/auth');
-	} else {
-		try {
-			await fetchUser(userToken);
-		} catch (e) {
-			throw redirect(307, '/auth');
-		}
-	}
-	return {};
-};
-
-export const ssr = false;
