+++ src/routes/(home)/+page.server.ts
@@ -1,27 +1,25 @@
-import { error, type Load } from '@sveltejs/kit';
-
 import { PUBLIC_USERS_API_URI } from '$env/static/public';
-import { token } from '../../../../stores/auth';
+import { error, type Load } from '@sveltejs/kit';
+import { token } from '../../stores/auth';
 
 export const load: Load = async () => {
 	let userToken;
-
 	token.subscribe((value) => {
 		userToken = value;
 	});
 
-	const res = await fetch(`${PUBLIC_USERS_API_URI}/users`, {
+	const res = await fetch(`${PUBLIC_USERS_API_URI}/teams`, {
 		method: 'GET',
 		headers: {
 			Authorization: 'Bearer ' + userToken
 		}
 	});
 
-	const { users } = await res.json();
+	const { teams } = await res.json();
 
 	if (res.ok) {
-		return { users };
+		return { teams };
 	}
 
-	throw error(404, 'Unable to fetch users');
+	throw error(404, 'Unable to fetch teams');
 };
