+++ src/routes/(home)/members/+page.server.ts
@@ -1,25 +1,27 @@
-import { PUBLIC_USERS_API_URI } from '$env/static/public';
 import { error, type Load } from '@sveltejs/kit';
+
+import { PUBLIC_USERS_API_URI } from '$env/static/public';
 import { token } from '../../../stores/auth';
 
 export const load: Load = async () => {
 	let userToken;
+
 	token.subscribe((value) => {
 		userToken = value;
 	});
 
-	const res = await fetch(`${PUBLIC_USERS_API_URI}/teams`, {
+	const res = await fetch(`${PUBLIC_USERS_API_URI}/users`, {
 		method: 'GET',
 		headers: {
 			Authorization: 'Bearer ' + userToken
 		}
 	});
 
-	const { teams } = await res.json();
+	const { users } = await res.json();
 
 	if (res.ok) {
-		return { teams };
+		return { users };
 	}
 
-	throw error(404, 'Unable to fetch teams');
+	throw error(404, 'Unable to fetch users');
 };
