+++ src/routes/(home)/projects/+layout.svelte
