+++ src/routes/(home)/projects/+page.svelte
