+++ src/routes/(home)/projects/directory/[projectName]/+page.svelte
