+++ src/routes/(home)/projects/directory/[projectName]/postings/[posting]/+page.server.ts
