+++ src/routes/(home)/teams/+page.svelte
@@ -58,7 +58,7 @@
 					</div>
 					<h3>{team.name}</h3>
 					<a
-						href="/portal/teams/{team}"
+						href="/teams/{team}"
 						class="bg-primary-500 rounded-lg px-4 py-2 text-white mt-4"
 					>
 						View Team
