+++ src/routes/(home)/teams/[team]/+page.svelte
