+++ src/routes/(home)/teams/[team]/+page.ts
